Overhauled blueprint operator and env variables
Environment variables were necessary to indicate to the operator its own
namespace. There are now passed in via helm chart or bash exports for
local dev.

We now also have most of the blueprint pipeline in place except the
final volume still needs inserting and a few checks for deletion are
also necessary.