Added sphinx docs outline
I want to add sphinx to document authentik-manager to make it easier for
everyone rather than having an extensive README. I realise this has
become one heck of a monorepo but for now this wont be an issue.

The docs are currently empty and there is no gitlab builder for them
yet. But that will be the next step.