Expanded cluster-roles perms to list and watch CM
Operator SDK insists on being able to list and watch configmaps cluster
wide. This is probably a quirk of being a global watching operator that
primarily deploys into one namespace, this is probably not the default
expected behaviour which is why it still likes to operate globally for
resources it should have under its wing. I have enabled only the minimal
necessary to get it to recognise that the configmaps do not in actual
fact belong to it, so then it can ignore them properly.