Expanded docs to include AKM rather than just AK
The docs was initially geared towards the original focus which was to
get a good install of authentik. Since we later had to add way more than
we originally intended, things have changed since then. Including the
now hybrid operator. This needed to be reflected in the docs. There is
still a lot more to add in particular to basic usage of AKM and
authentik.