Reduced number of operator pords to 1
The helm operator seems to conflict with itself. This could be that I
have misconfigured the operator such that it does not properly
co-operate on deployment of this resource. For now the easiest fix is to
reduce the number of pods, but this is not highly available. This
shouldnt be a large problem since we dont image it will change
frequently given the sensitivity, but at the same time we want to
maintain uptime. This will need further investigation to enable higher
availability.