Added annotation to allow sealed-secret overwrite
While this may break some installs early, this will be a very quick
indicator that some secret intervention is necessary. This should mean
people properly back up this important secret, and they dont only find
out later when they spin down the cluster and lose their secret for
whatever reason. This paired with the operator and it potentially not
creating a secret for the user by default should prevent this scenario
from ever needing to happen.