:bug: Added near complete fix for Raw abstraction
Raw abstraction was full functional however the use of it in the API
spec of AkBlueprints was incorrect. I have added tests and corrected the
usage. The tests check the API spec is marshal and unmarshallable using
base go-yaml which should mean it will work with k8s forks. I will add
further testing for k8s forks specifically but the kubebuilder API is
now working as intended.

Changelog: fixed