Added backchannel K8s fetching
This was necessary to attempt to avoid client-go REST from processing
the body. However the body was already wrong so this is a dead end.
Instead I am going to go for the original and inferior version, without
checking of the CRD, and just use a multiline string for these
infinitely frustrating custom YAML tags.